Moves extensions to be index.ts/index.tsx.
