+++ frontend/packages/markdown/src/extensions/details/index.tsx
