+++ frontend/packages/markdown/src/extensions/expandableImages/index.tsx
