+++ frontend/packages/markdown/src/extensions/index.ts
@@ -1,13 +1,13 @@
-export {callout} from './callout/callout';
-export {clickableText} from './clickableText/clickableText';
-export {details} from './details/details';
-export {embeds} from './embeds/embeds';
-export {expandableImages} from './expandableImages/expandableImages';
-export {externalLinks} from './externalLinks/externalLinks';
-export {inlineStyles} from './inlineStyles/inlineStyles';
-export {visualCodeBlock} from './visualCodeBlock/visualCodeBlock';
+export {callout} from './callout';
+export {clickableText} from './clickableText';
+export {details} from './details';
+export {embeds} from './embeds';
+export {expandableImages} from './expandableImages';
+export {externalLinks} from './externalLinks';
+export {inlineStyles} from './inlineStyles';
+export {visualCodeBlock} from './visualCodeBlock';
 export {
   vocabularyDefinition,
   type VocabularyLookup,
   type VocabularyTerm,
-} from './vocabularyDefinition/vocabularyDefinition';
+} from './vocabularyDefinition';
