+++ frontend/packages/markdown/src/extensions/vocabularyDefinition/index.tsx
